fix(adapter-planetscale): allow connecting without a database name
Currently if a database name is not provided, an empty string is passed
to the query engine or the query compiler, leading to broken queries:

```sql
SELECT ``.`User`.`id`, ``.`User`.`email`, ``.`User`.`status`, ``.`User`.`region`,
  ...
```

If there's no database name in the connection string, the `schemaName`
property in `ConnectionInfo` should not be set at all.

This is important when using sharded databases on PlanetScale.

Ref: https://github.com/prisma/prisma-engines/pull/5472